refactored to have more observables regarding trace operations and remove pause
